Extract backing up and restoring of database to a separate class in benchmark tests
